requirements.txt
django>=3.2,<3.3
gunicorn
psycopg2-binary
django-dotenv
python-dotenv
django-htmx
django-storages
boto3